+++ modules/ccv2/core/src/sap/commerce/toolset/ccv2/CCv2Constants.kt
@@ -40,6 +40,7 @@ object CCv2Constants {
     )
 
     val CLOUD_EXTENSIONS = arrayOf(
+        "azurehac",
         "azurecloudhotfolder",
         "cloudmediaconversion",
         "cloudcommons",
@@ -49,6 +50,7 @@ object CCv2Constants {
         "cloudstorefulfilmentprocess",
         "cloudstorefacades",
         "cloudstorecore",
-        "cloudstorecockpits"
+        "cloudstorecockpits",
+        "governor",
     )
 }