+++ modules/project/core/src/sap/commerce/toolset/project/codeInsight/daemon/EiLineMarkerProvider.kt
@@ -23,7 +23,7 @@ import com.intellij.psi.xml.XmlFile
 import com.intellij.util.xml.DomManager
 import sap.commerce.toolset.extensioninfo.model.ExtensionInfo
 
-abstract class EiSLineMarkerProvider : ExtensionLineMarkerProvider() {
+abstract class EiLineMarkerProvider : ExtensionLineMarkerProvider() {
 
     final override fun canProcess(psi: PsiFile) = (psi as? XmlFile)
         ?.let { DomManager.getDomManager(psi.project).getFileElement(it, ExtensionInfo::class.java) != null }